Chainstate helper class.
This adds the Chainstate class, which keeps track of the known
blocks and their tree structure in an SQLite database.  It provides
basic methods to determine fork points for branches and keep the
database up-to-date as new best tips get announced.

It also keeps track of move data for blocks that are not on the main
chain, so we can provide it to GSPs in case of reorgs for detaching
blocks even if the base blockchain does not allow us to get move data
for orphan blocks.