Syncing between chainstate and base chain.
This implements a helper class which runs a background task that
continuously tries to sync a Chainstate instance to whatever is the
latest state of the base chain.

It takes care of the initial sync when catching up to the tip, of
reorgs as they happen, and of course just keeps the tip updated during
normal operation as blocks come in.