Handle "short" reorgs in Sync.
This fixes a bug in Sync, which would fail to notice changes to the
tip when the new tip was on a lower height than the old one (i.e.
a reorg to a shorter chain).

Due to querying only for blocks from one beyond the tip, we would not
notice any different blocks in that case.  The fix is to always include
the current tip in the query range, so that we can detect any change to it.

We make explicitly sure to not fire the tip-updated callback in the case
that we just have the old tip, though.