Start with RPC server implementation.
Add a basic RPC server with the Xaya Core interface to the Controller.
It does not yet support all the methods as used by GSPs, but it does
support the main ones and in particular game_sendupdates.  This allows
us to also test this logic (requested notifications) in the Controller
with new unit tests.