Avoid deadlock with Controller shutdown.
Do not lock the Controller mutex for checking the santiy-check and
pruning variables during the tip-update, as this might lead to a
deadlock with the sync mutex during stopping of the Controller.

Instead just prevent them from being changed at all while the controller
is running, which should not be necessary anyway.