Add rngseed to BlockData.
Add the rngseed field to BlockData directly, and support it explicitly
in the Chainstate and ZmqPub classes.  The rngseed is a field that
libxayagame itself expects and reads, so it is not "general metadata".
By having it explicitly as part of BlockData, we emphasise to chain
connectors that it is necessary.