Keep holding chain mutex for TipUpdatedFrom.
When Sync calls the TipUpdatedFrom callback, retain the lock on the
chain mutex.  The callback itself will want to process the chain anyway,
so by just keeping hold of the mutex we ensure that the callback will
see a consistent view of the chainstate compared to e.g. the attached
blocks from Sync.