Optionally query for logs in a height range.
For blocks that are far behind the current tip, use a single eth_getLogs
call with a height range instead of individual calls by block hash
when retrieving moves from Ethereum logs.

While this method is not guaranteed to be safe when reorgs happen, it is
faster, and thus suitable for far-behind blocks during sync.

This fixes https://github.com/xaya/xayax/issues/1.