Track pending moves only optionally.
Instead of tracking pending moves always (as part of Start for a BaseChain
implementation), we have now a new method EnablePending.  Only when
requested by the main program and when the BaseChain signals it supports
pending moves will we enable the support and signal it through
getzmqnotifications.

With this, it will be optional to turn on or off pending tracking as
required by some GSP / Xaya X instance (especially when tracking of
pending moves may be expensive as with Ethereum).