Properly order pending notifications.
This introduces more logic around how pending moves are handled between
receiving them from the BaseChain and pushing them to ZMQ.  In particular,
we want to avoid a situation where the chainstate tip is already further
than the BaseChain notifications, and where we receive and then push
pendings while they are in fact already confirmed in what the Chainstate's
latest tip is.

In this change, we explicitly compare the latest TipChanged notification
tip with the current Chainstate tip, and use that information to drop,
relay or queue (and then relay later) pending transactions as they come in.

With this change, it may happen in rare cases that pendings that should be
relayed are dropped, but this is fine as pendings are best-effort only.  But
it will be ensured that any pendings that get pushed are correctly ordered
with respect to the block attach/detach notifications also sent.