Compile test contract with Truffle.
Instead of hard-coding and compiling the MultiMover test contract
from Python, set up a Truffle folder and use Truffle to compile it.
We then load and deploy the compiled artefact in the Python test.

This decouples the code a bit and also makes it easier to use the
MultiMover contract in other tests in the future.  But mostly this change
sets up general infrastructure for using Truffle with helper contracts
in Xaya X, which we will make use of for the state overlays that we need
to track pending moves.