Generate constants with Python script.
Use a Python script (based on the compiled Solidity artefacts and Web3.py)
to generate the Move event topic string rather than do so on the fly
from C++.  This makes the C++ code simpler, and also reduces our
dependency on the Ethereum JSON-RPC (removing the need for
web3_sha3).

This also lays the foundation for future changes where we also need
to introduce the overlay contract bytecodes into the C++ code.