Whitelist of tracked contracts for pending.
This introduces a whitelist of contract addresses, where only
transactions to a whitelisted address actually gets simulated with
eth_call to look for pending moves.

While this might miss some moves (e.g. from custom-deployed contracts),
it ensures that the large majority of transactions on the network, which
are not related to Xaya at all, will get quickly ignored.