Implement mempool retrieval for EVM chains.
This adds support for retrieving the mempool content on EVM chains
(as is needed for "getrawmempool", which in turn is used as part of
pending tracking by libxayagame).

We could have used txpool_content to directly mimic Xaya Core's RPC,
but this has two drawbacks:  First it is a geth extension and one that
is not readily available for some interfaces (and not even enabled by
default in geth), and second this is highly inefficient if we expect
most transactions to be unrelated to Xaya.

Thus instead, we keep track of our own "mempool" based on transactions
received from notifications, and we use a batched eth_getTransactionByHash
call to synchronise this with the underlying node whenever needed.