Make pruning mandatory.
This changes EnablePruning() on the Controller to SetMaxReorgDepth(), and
makes it mandatory to configure a depth (which in turn means that pruning
will have to be always enabled now).

In the future, we will remove the initial syncing step, and instead just
sync the last few blocks as per the max reorg depth to make running Xaya X
a lot more efficient.  This is a first preparatory step in this direction.