Completely remove pruned blocks.
This modifies how pruning works in the Chainstate:  Blocks that get
pruned are now completely removed, instead of just removing some extra
data and keeping the "block header".

The Chainstate class itself is also updated to support this, e.g. when
requesting a fork branch that ends in a pruned block.

Note that for now the Controller and Sync classes are not yet updated, and
in particular will lead to failures if a GSP requests blocks that have
been pruned.  This is something that will be properly handled in the future
(where we will just assume blocks not in the Chainstate are pruned, and
query the base chain for them).