Support completely pruned mainchain blocks.
This modifies the Controller and Sync classes to properly handle the
situation of fully pruned blocks on the main chain (e.g. when a GSP
requests updates from a block that has been removed entirely from the
Chainstate).

Also the compatibility RPCs getblockhash and getblockheader will
support requests for mainchain blocks that have been pruned in the
local chainstate.

To support this, the base chain gets a new method for checking if a given
block hash is on the main chain, and if so, at which height.  When an
unknown block is found (e.g. as 'from' for a game_sendupdates call), we
can defer to this method to handle the situation of a "valid" but pruned
block on the main chain; we get all the data for those blocks from the
base chain anyway already.