Allow reimporting of tips.
Instead of completely resetting the chainstate database when a new
"genesis" block is imported, allow just importing any new tips and
simply adding them to the database.  The only requirement is that they
must be a descendant of the current tip (which cannot be verified completely
and must be guaranteed by the caller).

With stronger pruning and "fast sync", we will get rid of the "genesis"
concept, and instead just work towards adding current tip - pruning height
as start block whenever Xaya X is restarted after some time off.  This is
a preparation step for that.