Remove 'genesis' block concept.
This completely removes the concept of a "genesis block" for the Xaya X
sync.  Instead of that, we always just use the maximum reorg depth
configured to sync the last couple of blocks from the base-chain tip
if started with an empty database.

This greatly speeds up initial sync, avoiding the unnecessary task of
processing all the past blocks through Xaya X.  When GSPs request ancient
blocks, we will just query them directly from the base chain anyway.