Integration test started/stopped Xaya X instance.
This extends the pruning.py integration test for Xaya X on Xaya Core
to include a second instance, which is started later on, stopped,
and started again to catch up to the underlying base chain (without
stying in sync).

We then verify that it correctly syncs to the expected tips, and also
keeps the old branch available for requesting blocks even with pruning
turned on.  This is particularly useful to ensure that the "fast sync"
logic works.

The new test is applied only on Xaya X on Xaya Core, as it tests the
underlying main logic and not something base-chain specific.