Fast catch-up based on pruning depth.
This implements "fast catch-up":  If we already have blocks in the
chainstate but the old tip is far behind the base-chain tip, then
we re-import the current base-chain tip instead of processing all missed
blocks (based on the pruning depth similar to how the initial block
would be imported).

Any branch the original chainstate might be synced to is resolved
first, though, so that we will always be able to properly serve blocks
requested by GSPs that might be at a fork.