Implement RPC proxy between Xaya X and Core.
This implements a basic proxy class for JSON-RPC, which supports
artificially disconnecting it or triggering errors (without affecting
the underlying actual RPC server that backs it).

In the test framework for Xaya X on Core, we add that proxy in between
Xaya X and Xaya Core.  For now, this is just a no-op, but in the future it
will allow us to test situations where the base chain (Xaya Core in this
case) throws temporary errors to Xaya X.