Fix flakiness in pruning tests.
The pruning tests were flaky, because they explicitly expected all
attach and detach messages to be received from Xaya X when the chain
was being modified (e.g. with a reorg).

With fast catch-up, this may no longer be the case, since Xaya X could just
do a fast catch-up instead if there are some random delays and it does not
stay completely up-to-date all the time.

To fix this, we just wait for the final block-attach ZMQ message, which
is good enough.  After all, these tests are about verifying explicitly
requested ZMQ messages and RPCs that inspect the chain state, not about the
ZMQs received during syncing.