Gracefully handle base-chain errors.
Explicitly allow the BaseChain methods to throw exceptions (e.g. because
the underlying RPC connection is broken somehow), and implement
graceful degradation where possible in the Xaya X main code when
that happens.

For instance, Sync will now be able to tolerate temporary base-chain errors,
and will just recover when the errors stop.  RPC methods that call through
to the base-chain will catch and rethrow the errors that appear (so the
final RPC client will receive them as JSON-RPC internal errors).