Update Ganache invokation.
Update the way Ganache is run in tests.  For one, the command is changed
to "ganache" instead of "ganache-cli", which matches changes made
to Ganache itself.

Secondly, we explicitly turn on --chain.vmErrorsOnRPCResponse, which
ensures that eth_call RPCs throw when the tx reverts, which also matches
what geth does.  This is the default already for some versions of Ganache,
but explicitly specifying the flag can't hurt in any case.