Add "mvid" for EVM chains.
On EVM chains, the txid is not a good identifier for moves, as a
single transaction can trigger multiple moves, and also e.g. the
same transaction executed in different contexts (perhaps due to a reorg)
can trigger different (even a different number of) moves.

To help games work around this, for instance to construct things like
channel IDs and reinits, this introduces a new "mvid" concept and field
that gets returned with moves in the EthChain implementation.

The "mvid" is the Keccak hash of the ABI-encoded data for the
move event.  As such, it commits to, in particular, the name, nonce
(making sure it is unique except for reorgs) as well as key attributes
like the move string and WCHI payments.  This overall has then very similar
properties to the txid on Xaya Core.