Commit to chain ID in EVM signatures.
For VerifyMessage on EVM chains, explicitly include the chain ID in the
signature scheme used.  This ensures that replay attacks between different
chains will not be possible for Xaya signatures.

On the other hand, this breaks "compatibility" between the Xaya signing
scheme and arbitrary EVM signatures, but that's not a feature we need,
since Xaya signatures (e.g. for XID) are always a specific usecase.