Consistently use base64 for verifymessage.
With this update, we clarify that the signature for the RPC verifymessage
should be passed as base64 (how it is done already for Xaya Core, but it
will be the same for other chains as well).  The base chain implementation
of VerifyMessage on the other hand is based on raw bytes (which can be
encoded in a chain-specific way before processing it with chain-specific
RPC endpoints or tools).