Allow optional toblock in game_sendupdates.
This extends the game_sendupdates method in the RPC server to accept
an optional "toblock" argument.  The argument must be empty for now
(i.e. is not yet actually handled).

libjson-rpc-cpp does not directly support method overloading or optional
arguments, so this by itself (just making the server accept it) is
what this change is about.

This is done by overriding the method handling calls in the RPC server,
and special-casing game_sendupdates.  In addition, by declaring the
method without arguments in the spec file, we turn off argument validation
on the libjson-rpc-cpp side, and do our own validation in the overridden
handler method.