Clarify forkHeight in PushZmqBlocks.
Previously, the internal forkHeight variable set inside the PushZmqBlocks
logic was set to the height of the first block to attach from the main
chain.  Now it is one less, so that it actually is the "fork point" (i.e.
the block on the main chain on which the fork is based, or the from
block itself if it is on the main chain).  We also set explicitly the
forkPoint variable to its block hash.

This makes the code easier and more intuitive, and also simplifies it a
bit later on (by means of the new forkPoint).