Support 'to' block in game_sendupdates.
This adds actual support for an explicit 'to' block to game_sendupdates,
as long as the target block is on the main chain (but it can be an old
block instead of the tip).  This is useful in combination libxayagame's
upcoming support for general "target blocks", see
https://github.com/xaya/libxayagame/issues/130.

Implements https://github.com/xaya/xayax/issues/7.