Use serialised BlockData internally.
When storing blocks internally (for active branches that are not yet
pruned), serialise BlockData to a BLOB and store that instead of
separately storing the various fields and even moves in their own
table.

This simplifies the code, and also makes explicit in the database schema
what type of data is needed internally by Xaya X to manage the chain state,
and what is just passed on to GSPs.