Implement block-cache logic.
This adds logic (in form of a new BaseChain implementation that can
sit on top of another BaseChain) for caching "old" blocks.  This means
that when GetBlockRange is called for a range far behind the current tip,
which is assumed to be final, then the blocks retrieved from the underlying
base chain can be cached into some storage, and future calls will
retrieve the blocks from there instead of querying for them again.

With this, load from the base-chain client can be moved onto e.g.
a database server, which may help to scale the system.