Add memory block cache to xayax-eth.
This adds a new --blockcache_memory boolean flag to xayax-eth, which
when set will add an in-memory block cache on top of the EVM base chain.
A new integration test turns on that flag and checks that it works
(in a very basic fashion).

Note that a memory block cache is typically not very useful, except
for testing the block cache logic in general.  So this has likely no
impact or use in production (but allows for a basic integration test).