Add "depth" for tracked games.
Instead of just keeping a list of game IDs that are tracked (and removing
them when untracked), store a "depth" value for each game.  This is
incremented when the game is tracked, and decremented when it is untracked.
With this, it is possible to connect multiple GSPs for a single game ID
to one Xaya X instance, and they won't disturb each others' tracked games.