Switch from Truffle to Foundry.
This replaces Truffle to build the integrated smart contracts to
Foundry.  We use forge to build all the Solidity code, run the
unit tests (recoded in Solidity), and update various places that
access the build artefacts to the new path structure from Foundry.

Dependencies (in particular, the XayaAccounts contract) are included
now as git submodules, rather than via npm.

The Docker build is updated, too.  Since it seems very hard to install
Foundry in Alpine (using prebuilt binaries that depend on glibc), we
switch the Docker build to be based on Ubuntu.